improve unit testing case for package gdb
