+++ database/gdb/gdb_z_mysql_basic_test.go
@@ -7,10 +7,10 @@
 package gdb_test
 
 import (
-	"testing"
-
+	"github.com/go-sql-driver/mysql"
 	"github.com/gogf/gf/database/gdb"
 	"github.com/gogf/gf/test/gtest"
+	"testing"
 )
 
 func Test_Instance(t *testing.T) {
@@ -25,5 +25,18 @@ func Test_Instance(t *testing.T) {
 		err2 := db.PingSlave()
 		t.Assert(err1, nil)
 		t.Assert(err2, nil)
+	})
+}
+
+// Fix issue: https://github.com/gogf/gf/issues/819
+func Test_Func_ConvertDataForTableRecord(t *testing.T) {
+	type Test struct {
+		ResetPasswordTokenAt mysql.NullTime `orm:"reset_password_token_at"`
+	}
+	gtest.C(t, func(t *gtest.T) {
+		m := gdb.ConvertDataForTableRecord(new(Test))
+		t.Assert(len(m), 1)
+		t.AssertNE(m["reset_password_token_at"], nil)
+		t.Assert(m["reset_password_token_at"], new(mysql.NullTime))
 	})
 }
